Fixed logout link (#131).
