+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed download button in the File Manager (#121).
+- Fixed logout link (#131).
 
 ## [0.9.4] - 2023-10-20
